dma-buf: samsung_heap: align the types of fd_flags and heap_flags (v6.11 compatibility)
Upstream commit [1] causes downstream pixel drivers to fail:

private/google-modules/soc/gs/drivers/dma-buf/heaps/samsung/system_heap.c:193:14: error: incompatible function pointer types initializing 'struct dma_buf *(*)(struct dma_heap *, unsigned long, u32, u64)' (aka 'struct dma_buf *(*)(struct dma_heap *, unsigned long, unsigned int, unsigned long long)') with an expression of type 'struct dma_buf *(struct dma_heap *, unsigned long, unsigned long, unsigned long)' [-Wincompatible-function-pointer-types]
  193 |         .allocate = system_heap_allocate,
      |                     ^~~~~~~~~~~~~~~~~~~~

Align the types making sure the children use the correct type too and no
casting to a smaller type occurs. In fact only samsung_export_dmabuf()
needed to be updated. This method sets ```exp_info.flags = fd_flags;```
where an implicit cast from u32 to int happens. All shall be fine.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=b9578c49456340ca4d3c7ddbaca054ffc2b51bc1 [1]
Test: TH
Bug: 359443502
(cherry picked from commit 6f1c88d67ad83b2e530233d03092316b3d28e6b2)

+++ drivers/dma-buf/heaps/samsung/samsung-dma-heap.h
@@ -84,7 +84,8 @@ void samsung_dma_buffer_free(struct samsung_dma_buffer *buffer);
 int samsung_heap_add(struct device *dev, void *priv,
 		     void (*release)(struct samsung_dma_buffer *buffer),
 		     const struct dma_heap_ops *ops);
-struct dma_buf *samsung_export_dmabuf(struct samsung_dma_buffer *buffer, unsigned long fd_flags);
+struct dma_buf *samsung_export_dmabuf(struct samsung_dma_buffer *buffer,
+				      u32 fd_flags);
 void samsung_track_buffer_destroyed(struct samsung_dma_buffer *buffer);
 
 #define DMA_HEAP_VIDEO_PADDING (512)
